Moving the error module out of the utils directory
