+++ src/drm/base-drm-protocol.js
@@ -1,7 +1,7 @@
 // @flow
 /* eslint-disable no-unused-vars */
 import getLogger from '../utils/logger'
-import Error from '../utils/error/player-error'
+import Error from '../error/player-error'
 import DrmSupport from './drm-support'
 import {DrmScheme} from './drm-scheme'
