+++ src/engines/html5/media-source/base-media-source-adapter.js
@@ -2,7 +2,7 @@
 /* eslint-disable no-unused-vars */
 import FakeEvent from '../../../event/fake-event'
 import FakeEventTarget from '../../../event/fake-event-target'
-import Error from '../../../utils/error/player-error'
+import Error from '../../../error/player-error'
 import {CUSTOM_EVENTS} from '../../../event/events'
 import getLogger from '../../../utils/logger'
 import Track from '../../../track/track'
