+++ src/error/severity.js
